GH-3 create resource skeletons
Closes #3

Add resource skeleton for both resources.  Skeleton includes schema,
js source files, and rpdk config.  The schema outlines what the resource
looks like in cloudformation.  The js files become a lambda that executes
on the resource events.  The rpdk config tells cloudformation about how
to create the lambda.

Add delivery and infra cloudformation templates.  The delivery file adds
resources needed for the deployment and delivery of the project.  The
infra template contains the resources for the deployed resource.

Add deploy script and build commands.  Still need to modify this script to
deploy the infra yml file instead of doing it directly.